infra/docker_images_sync/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
boto3==1.26.89
    # via -r requirements.in
botocore==1.29.89
    # via
    #   boto3
    #   s3transfer
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.2
    # via botocore
pyyaml==6.0
    # via -r requirements.in
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.15
    # via botocore
